requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
